Completed Tutorial Scene
I added the radio into the scene it was giving me trouble before. I added a person model, but I am unsure if this is where the starting location will be. The goal is to follow the cones from the model to reach the table, where the user is able to equip the iPad and radio. Take a picture of the truck with the iPad, and interact with the radio.